+++ lib/td_qx/quality_controls/control_properties/ratio.ex
@@ -6,6 +6,7 @@ defmodule TdQx.QualityControls.ControlProperties.Ratio do
   use Ecto.Schema
 
   import Ecto.Changeset
+  alias TdQx.DataViews.Queryable
   alias TdQx.DataViews.Resource
   alias TdQx.Expressions.Clause
 
@@ -14,12 +15,14 @@ defmodule TdQx.QualityControls.ControlProperties.Ratio do
   embedded_schema do
     embeds_one(:resource, Resource, on_replace: :delete)
     embeds_many(:validation, Clause, on_replace: :delete)
+    embeds_one(:segmentation, Queryable, on_replace: :delete)
   end
 
   def to_json(%__MODULE__{} = ratio) do
     %{
       resource: Resource.to_json(ratio.resource),
-      validation: Clause.to_json(ratio.validation)
+      validation: Clause.to_json(ratio.validation),
+      segmentation: Queryable.to_json(ratio.segmentation)
     }
   end
 
@@ -30,6 +33,23 @@ defmodule TdQx.QualityControls.ControlProperties.Ratio do
     |> cast(params, [])
     |> cast_embed(:resource, with: &Resource.changeset/2, required: true)
     |> cast_embed(:validation, with: &Clause.changeset/2, required: true)
+    |> cast_embed(:segmentation, with: &segmentation_changeset/2)
+  end
+
+  defp segmentation_changeset(%Queryable{} = struct, %{} = params) do
+    struct
+    |> Queryable.changeset(params)
+    |> validate_change(:type, fn _, type ->
+      if type == "group_by" do
+        []
+      else
+        [{:type, "segmentation queryable must be of type 'group_by'"}]
+      end
+    end)
+  end
+
+  defp segmentation_changeset(_struct, params) do
+    segmentation_changeset(%Queryable{}, params)
   end
 
   def enrich_resources(%__MODULE__{} = ratio, enrich_fun) do